fix lint errors, move unit test into test folders
