+++ src/components/__tests__/.eslintrc
@@ -1,0 +1,11 @@
+{
+    "extends" : "../../../.eslintrc.js",
+    "env"     : {
+        "mocha" : true
+    },
+    "globals" : {
+        "expect" : false,
+        "should" : false,
+        "sinon"  : false
+    }
+}
