+++ src/container/__tests__/UserRepo.test.js
@@ -1,22 +1,23 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-
-import User from './User';
 import renderer from 'react-test-renderer';
 
-describe('User component', () => {
+import UserRepo from '../UserRepo';
+
+const user = {};
+const repo = { branches: [] };
+const github = { user, repo };
+
+describe('UserRepo component', () => {
     it('should render without crashing', () => {
         const div = document.createElement('div');
-        ReactDOM.render(<User />, div);
+        ReactDOM.render(<UserRepo />, div);
     });
 
     it('should render as expected', () => {
-        const props = {
-            avatar_url: 'www.avatar.url',
-            login: 'login',
-            company: 'comp',
-        };
-        const appContainer = renderer.create(<User {...props} />).toJSON();
+        const props = { github };
+
+        const appContainer = renderer.create(<UserRepo {...props} />).toJSON();
         expect(appContainer).toMatchSnapshot();
     });
 });
